cql3: statement_restrictions: split _partition_range_restrictions into three cases
_partition_range_restrictions are a vector of expressions, one per
partition key column, except that it can be empty if there is no
restriction on the partition that can be translated to a read command,
and if the restriction is on a token range, the first element only
is used.

Separate the three cases into distinct structs. After this, additional
work can be done utilizing the specialization.