:test.py: integrate python tests to be executed with pytest runner
This will move responsibility for running tests with pytest in the same
manner as it was done with boost tests. From this commit, test.py is not
responsible anymore for running python tests and relies completely on
pytest.
This is another step for unification of test execution.
Use xdist directly, since it's already a prerequisite for the test.py
Remove `-s`, so print inside the test will not be thrown to the terminal

NOTE: this is a breaking change. From this commit, several directories
with tests will require a path to the file to launch the test.
Affected directories
test/alternator
test/broadcast_tables
test/cql
test/cqlpy
test/rest_api